Update issue 217 SASL:   Fixed incorrect mask in cg_service::emit_tex_grad_impl.   Fixed all pixel shader external intrinsic signature.   pixel shader unit works.   C++-Implementation of external intrinsic works.   All test cases works.

+++ salviar/include/sampler_api.h
@@ -19,18 +19,18 @@ extern "C"
 
 	// Packaged version
 	void salviar_tex2Dgrad_pkg(
-		eflib::vec4* results, uint16_t mask,
+		eflib::vec4* results, uint32_t mask,
 		salviar::sampler* samp, eflib::vec2* coords,
 		eflib::vec2 const* ddxs, eflib::vec2 const* ddys );
 	void salviar_tex2Dbias_pkg(
-		eflib::vec4* results, uint16_t mask,
+		eflib::vec4* results, uint32_t mask,
 		salviar::sampler* samp, eflib::vec4* coords,
 		eflib::vec2 const* ddxs, eflib::vec2 const* ddys );
 	void salviar_tex2Dlod_pkg (
-		eflib::vec4* results, uint16_t mask,
+		eflib::vec4* results, uint32_t mask,
 		salviar::sampler* samp, eflib::vec4* coords );
 	void salviar_tex2Dproj_pkg(
-		eflib::vec4* results, uint16_t mask,
+		eflib::vec4* results, uint32_t mask,
 		salviar::sampler* samp, eflib::vec4* coords,
 		eflib::vec4 const* ddxs, eflib::vec4 const* ddys );
 }
